% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_data.R
\name{analyze_data}
\alias{analyze_data}
\title{analyze_data}
\usage{
analyze_data(inp, time_col = "week", pattern = "aldi|rewe|edeka",
  pattern_col = "AuftraggeberBegunstigter")
}
\arguments{
\item{inp}{input data.frame as provided by import data}

\item{time_col}{col.name of inp which contains temporal data}

\item{pattern}{used to subset the data, multiple input separated by "|"}

\item{pattern_col}{column to pattern should be applied}
}
\value{
dataframe with aggregated data
}
\description{
analyze_data
}
\examples{
dd <- import_data()

#' Example 1: Defaults --------------------------------------------------
dd2 <- analyze_data(dd)

ggplot(dd2, aes(x = time, y = value, group = variable, col = variable),
       alpha = .6) +
  geom_line(lwd = 1) +
  geom_point(size = 3) +
  theme_dark() +
  theme(axis.text.x = element_text(angle = 90, hjust = .5))


#' Example 2: Lohn --------------------------------------------------

#'remove first + last month from data
dd_ <- subset(dd,!month \%in\% range(dd$month))

dd2 <- analyze_data(dd_,pattern = "gfk|fraunhofer", time_col = "month")


ggplot(dd2, aes(x = time, y = value, group = variable, col = variable),
       alpha = .6) +
  geom_line(lwd = 1) +
  geom_point(size = 3) +
  geom_smooth(method = "lm", lwd = .5, lty = "dashed") +
  theme_dark() +
  theme(axis.text.x = element_text(angle = 90, hjust = .5))


#' Example 3: Vodafone -----------------------------------------------
dd2 <- analyze_data(dd_,pattern = "vodafone", time_col = "month")


ggplot(dd2, aes(x = time, y = value, group = variable, col = variable),
       alpha = .6) +
  geom_line(lwd = 1) +
  geom_point(size = 3) +
  geom_smooth(method = "lm", lwd = .5, lty = "dashed") +
  theme_dark() +
  theme(axis.text.x = element_text(angle = 90, hjust = .5))


#' Example 4: Hausgeld ---------------------------------------------
dd2 <- analyze_data(dd_,pattern = "FRIEDENSTRASSE", time_col = "month")



ggplot(dd2, aes(x = time, y = value, group = variable, col = variable),
       alpha = .6) +
  geom_line(lwd = 1) +
  geom_point(size = 3) +
  geom_smooth(method = "lm", lwd = .5, lty = "dashed") +
  theme_dark() +
  theme(axis.text.x = element_text(angle = 90, hjust = .5))
}

