% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data}
\alias{import_data}
\title{import_data}
\usage{
import_data(last_file = get_last_file())
}
\arguments{
\item{last_file}{character file giving path to csv-input}
}
\value{
data.frame with imported and cleaned data
}
\description{
import_data
}
\details{
if last_file is not provided, the last downloaded file is automatically chosen
}
\examples{
import_data()
}

